Fix compile errors from GCC.
How did this ever compile as it was???

+++ test/LexerTest.cc
@@ -15,7 +15,7 @@ tokenize(string input)
     string output;
     int eofcount = 0;
 
-    for (int i = 0; i < tokens.size(); ++i) {
+    for (size_t i = 0; i < tokens.size(); ++i) {
         if (tokens[i].getTokenType() == tok_eof) {
             eofcount++;
         }
